llm_hpo_project_definitivo/requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
datasets==4.4.1
deap==1.4.3
dill==0.4.0
exceptiongroup==1.3.0
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.2.1
numpy==1.26.3
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
peft==0.17.1
pillow==11.3.0
propcache==0.4.1
psutil==7.1.3
pyarrow==21.0.0
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
